Move temporal_degrain
Changed my mind :tehe:
